Avoid body parsing for requests with no Content-Type and no body
Several issues have been opened in the past because by default, Fastify responds with an error to POST requests without a body. People expect requests without a body to work without having to set a "catch all" content type parser. It's counter-intuitive to have to set a body parser to be able to handle requests without a body.

This change avoids running any content type parser if the request has no `Content-Type` header and no body. From RFC 7230 Section 3.3, we can determine if a request has a body by checking the `Transfer-Encoding` and `Content-Type` headers.