Update Server.md (#1718)
In the genReqId sention, it is mentioned in note, that the provided function for it will not be called if the "request-id" header is available. Hence, I think in the mentioned example, it should only return "i++" instead of "req.headers['request-id'] || i++"' as here the header here will not be present always.